Updated paths to use English
We don't translate URL paths. So `/en/patrons/join` in French is `/fr/patrons/join`, instead of something like `/fr/mecenes/rejoindre`.

If every language existed on its own, we could translate the URLs too. However, when a language is not available, we fallback to English. But that requires that we know what the English page is. Now, we can just replace the language prefix, but if the URL itself is also translated, that becomes a whole new level of complexity.

If we are going to do this, it will require changes to the code first.