:fire: Removed Markdown file that was in the wrong place
This was a bit of a rabbit hole to figure out, because it happened
a bunch of commits ago, and I didn't realize at the time. But here's
what happened.

A file fr.md was save (by accident) in the mardown/showcase/2017
folder. It's not supposed to go there, as each showcase lives in
it's own folder one level deeper, but mistakes happen.

But there's a twist: The 'img' field in the frontmatter always
refers to an image file in the same folder. Gatsby will see that
and use Sharp to turn that field into a sharp object.
Now, we can use the img field in the frontmatter as a Sharp object.

However, by dropping this markdown file in the wrong place, its
img field points to a file that isn't there (as only the .md file
was misplaced). Since there is no such file, the img field is now
just treated as a string. And since all fields need to be of the
same type throughout the graph, all of a sudden, all img fields in
the frontmatter are simple string values. Which of course causes
all sort of problems, and the build will fail early as the
GraphQL queries won't run because they want to query subfields
of a simple string value, and that's not possible.

So next time you see an error like:

> Field "img" must not have a selection since type "String" has no
subfields.

It means that the img frontmatter field is pointing to a file that
does not exist.

In the future, we should run tests to catch this early.